Update mutation.py
'MUTATION_PROBABILITY' and 'MUTATION_EVENTS' are set by default. Therefore the Exception can never occur and I removed it. 

Futhermore, the code as it was lead to a bug.

The bug can be reproduced by only setting 'MUTATION_PROBABILITY' = 0. as a parameter. Expected: No mutations. What happens in the original code is that params['MUTATION_EVENTS'] defaults to 1 and the condition 'MUTATION_PROBABILITY' evaluates to true by default. The else part of the condition is evaluated and p_mut  :=1/eff_length.
Thus, with 'MUTATION_PROBABILITY' == 0. p_mut will be > 0. in some cases.